Yay! Header getters can take &self, not &mut self!
This is achieved by using mucell to make the mutability safe. The
practical effect of it is that we’re using smart references (that is,
types implementing `Deref<T>`) rather than references (`&T`) and clones
(`T where T: Clone`). This is just a teensy bit less efficient all the
time but allows convenient use of `Cow` types so that for various cases
it will actually end up more efficient. But the fact that it doesn’t
need `&mut self` is the really big win. In all, I’m definitely happy.