Rust update.
There is a certain performance regression in `HeaderShowAdapter` because
of the new UTF-8 rigidity of the fmt infrastructure, so while
`fmt_header` works on the byte-based `io::Writer` there must
unfortunately be an intermediate allocation, rather than writing
directly to the formatter.

The `Header` trait has also had to be split with the `parse_header`
method shifted to a new `ToHeader` trait, because having no receiver it
is not object-safe and so without the change a `Box<Header>` object
could not be created any more.

The rest is comparatively mechanical.